pyproject.toml
[project]
name = "visual-question-answering"
version = "0.1.0"
description = "A modular framework for fine-tuning a Visual Question Answering on the VizWiz dataset"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Al-ameen007" },
]

# ── Core dependencies ──────────────────────────────────────────────
dependencies = [
    "transformers>=4.30.0",
    "torch>=2.0.0",
    "tqdm>=4.60",
    "requests>=2.28",
    "Pillow>=9.0.0",
]

# ── Optional dependency groups ─────────────────────────────────────
[project.optional-dependencies]
# For Finetuning Base Model
train = [
    "datasets>=2.0.0",
]
# For Visualizing on inference
viz = [
    "matplotlib>=3.5.0",
]
# Everything (for development convenience)
all = [
    "visual-question-answering[train,viz]",
]
# Dev/test tools
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "nox>=2023.4",
    "ruff>=0.4",
    "black>=24.0",
    "mypy>=1.8",
]

# ── CLI entry points ──────────────────────────────────────────────
[project.scripts]
vqa-train          = "vqa.train:main"
vqa-infer          = "vqa.inference:main"

# ── Build system ──────────────────────────────────────────────────
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vqa"]

# ── Ruff (replaces flake8, isort, pyflakes, pycodestyle) ──────────
[tool.ruff]
src = ["src", "tests"]
line-length = 100

[tool.ruff.lint]
select = [
    "F",    # pyflakes
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "I",    # isort (import sorting)
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "A",    # flake8-builtins
    "SIM",  # flake8-simplify
]

[tool.ruff.lint.isort]
known-first-party = ["vqa"]

# ── Pytest ────────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
